Incoporate feedback from review.
